Nw-854 Refactored tracking event
