+++ lib/events/index.js
@@ -1,7 +1,7 @@
 // Module responsible for handling dom events and delegating the action elsewhere
-
 var domUtils = require("../utils/dom"),
 	StateControl = require("./states"),
+	tracking = require("./tracking"),
 	getContainerIds = require("../utils/dom/getContainerIds");
 
 var KeysDict = {
@@ -25,7 +25,7 @@ function tophatEvents( document, tophatElement ) {
 function attachTophatEvents( document, tophatElement ) {
 	domUtils.attachDomEvent( tophatElement, "click", domUtils.proxy( tophatElement, clickhandler ) );
 	domUtils.attachDomEvent( document, "click", domUtils.proxy( tophatElement, cancelhandler ) );
-	domUtils.attachDomEvent( tophatElement, "click", domUtils.proxy( tophatElement, trackingHandler ) );
+	domUtils.attachDomEvent( tophatElement, "click", domUtils.proxy( tophatElement, tracking.handler ) );
 	domUtils.attachDomEvent( tophatElement, "submit", domUtils.proxy( tophatElement, searchHandler ) );
 	domUtils.attachDomEvent( tophatElement, "keydown", domUtils.proxy( tophatElement, keyHandler ) );
 
@@ -34,7 +34,6 @@ function attachTophatEvents( document, tophatElement ) {
 		wait = Math.round(1000 / fps);
 	domUtils.attachDomEvent( window, "resize", domUtils.throttle(domUtils.proxy( tophatElement, resizeHandler), wait) );
 }
-
 
 // Handle key press to delegate action
 function keyHandler(e) {
@@ -93,59 +92,6 @@ function resizeHandler() {
 	this.state.setAriaStates();
 }
 
-function trackingHandler( ev ) {
-	var target = ev.target || ev.srcElement;
-
-	while ( !validateTrackedElement( target ) ) {
-
-		if ( ~target.className.indexOf( "nice-tophat" ) ) {
-			target = undefined;
-			break;
-		}
-
-		target = target.parentNode;
-	}
-
-	var href = target && ( target.href || target.getElementsByTagName("a")[0].href );
-
-	if (!target || !href) return;
-
-	var category = "TopHat";
-	var labelEl = ( domUtils.find( target, "menu-label" )[0] || target);
-	var action =  "Tophat click";
-	var label = (labelEl.textContent || labelEl.innerText || labelEl.innerHTML);
-
-	switch ( cleanClass( target.className ) ) {
-		case "menu-profile":
-			action = "Your Profile" + (~this.className.indexOf( target.className + "-open" ) ? " expanded" : " collapased");
-			href = false;
-			break;
-
-		case "menu-evidence":
-		case "menu-search":
-			action += (~this.className.indexOf( target.className + "-open" ) ? " expanded" : " collapased");
-			href = false;
-			break;
-	}
-
-	ev.preventDefault();
-
-	sendTrackedEvent( category, action, label, function() {
-		if (href) window.location.href = href;
-	});
-}
-
-function validateTrackedElement( target ) {
-	var isTrackedElement = target &&
-            target.nodeName.toLowerCase() === "li" &&
-            ( ~target.className.indexOf( "menu-" ) || ~target.className.indexOf( "evidence-" )  );
-
-	return !!isTrackedElement;
-}
-
-
-
-
 function clickhandler( ev ) {
 	var target = ev.target || ev.srcElement;
 
@@ -165,7 +111,7 @@ function clickhandler( ev ) {
 	ev.preventDefault();
 
 	(ev.target || ev.srcElement).blur();
-	var elementId = cleanClass( target.className );
+	var elementId = domUtils.removeClass( target, "active" );
 	switch (elementId) {
 		case "menu-evidence":
 			this.state.toggleEvidence();
@@ -242,72 +188,12 @@ function searchHandler( ev ) {
 		var term = ( q !== placeholder ) ? escape( q.replace(/\s/g, "+") ) : "";
 		var location = target.action.replace(/%(25)?term/ig, term);
 
-		sendTrackedEvent( "Search", term, location, function() {
+		tracking.sendTrackedEvent( "Search", term, location, function() {
 			window.location.href = location;
 		});
 
 		ev.preventDefault();
 	}
-}
-
-
-
-
-// Tracking helpers
-
-function sendTrackedEvent( category, action, label, cb ) {
-	if (cb) {
-		var timeout = window.setTimeout( cb, 2000 );
-	}
-
-	if ( window.dataLayer && typeof window.dataLayer.push === "function" ) {
-		sendDataLayerEvent( category, action, label, cb );
-	}
-
-	else if ( window._gaq && typeof window._gaq.push === "function" ) {
-		sendGAEvent( category, action, label );
-	}
-
-	else if ( typeof window.ga === "function" ) {
-		sendUAEvent( category, action, label );
-	}
-
-	window.clearTimeout(timeout);
-}
-
-function sendDataLayerEvent( category, action, label, callback) {
-	var data = {
-		event: "TopHat",
-		eventCategory: category,
-		eventAction: action,
-		eventLabel: label,
-		eventCallback: callback
-	};
-
-	window.dataLayer.push( data );
-}
-
-function sendGAEvent( category, action, label ) {
-	var data = [ "_trackEvent", category, action, label ];
-
-	window._gaq.push( data );
-}
-
-function sendUAEvent( category, action, label ) {
-	var data = {
-		category: category,
-		action: action,
-		label: label
-	};
-
-	window.ga( "send", "event", data );
-}
-
-
-// utils
-
-function cleanClass( className ) {
-	return className.replace(" active", "");
 }
 
 module.exports = tophatEvents;
