+++ lib/utils/dom.js
@@ -356,6 +356,16 @@ utils.htmlEncode = function htmlEncode( html ) {
 		document.createTextNode( html ) ).parentNode.innerHTML;
 };
 
+/**
+ * Remove the class from the specified element
+ *
+ * @param      {HTMLElement}  { The element to remove class from. }
+ * @return     {String}  { The class to be removed }
+ */
+utils.removeClass = function( element, className ) {
+	return element.className.replace(" " + className, "");
+};
+
 (function(){
 	if (!Array.prototype.indexOf)
 	{
